# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Hifdh Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  password: string;
}

export interface AuthResponse {
  authenticated: boolean;
}

export interface Student {
  id: number;
  name: string;
  currentSurah: number;
  currentAyah: number;
  startDate: string;
  notes?: string | null;
  active: boolean;
  createdAt: string;
}

export interface CreateStudentRequest {
  name: string;
  currentSurah: number;
  currentAyah: number;
  startDate: string;
  notes?: string | null;
}

export interface UpdateStudentRequest {
  name?: string;
  currentSurah?: number;
  currentAyah?: number;
  notes?: string | null;
  active?: boolean;
}

export type Grade = (typeof Grade)[keyof typeof Grade];

export const Grade = {
  excellent: "excellent",
  good: "good",
  needs_repeat: "needs_repeat",
  incomplete: "incomplete",
} as const;

export interface DailyEntry {
  id: number;
  studentId: number;
  date: string;
  newMemorizationFromSurah?: number | null;
  newMemorizationFromAyah?: number | null;
  newMemorizationToSurah?: number | null;
  newMemorizationToAyah?: number | null;
  newMemorizationCompleted: boolean;
  newMemorizationGrade?: Grade | null;
  rmvFromSurah?: number | null;
  rmvFromAyah?: number | null;
  rmvToSurah?: number | null;
  rmvToAyah?: number | null;
  rmvCompleted: boolean;
  rmvGrade?: Grade | null;
  reviewFromSurah?: number | null;
  reviewFromAyah?: number | null;
  reviewToSurah?: number | null;
  reviewToAyah?: number | null;
  reviewCompleted: boolean;
  reviewGrade?: Grade | null;
  extraReviewFromSurah?: number | null;
  extraReviewFromAyah?: number | null;
  extraReviewToSurah?: number | null;
  extraReviewToAyah?: number | null;
  teacherNotes?: string | null;
  daySuccessful: boolean;
  createdAt: string;
}

export interface UpsertEntryRequest {
  newMemorizationFromSurah?: number | null;
  newMemorizationFromAyah?: number | null;
  newMemorizationToSurah?: number | null;
  newMemorizationToAyah?: number | null;
  newMemorizationCompleted: boolean;
  newMemorizationGrade?: Grade | null;
  rmvFromSurah?: number | null;
  rmvFromAyah?: number | null;
  rmvToSurah?: number | null;
  rmvToAyah?: number | null;
  rmvCompleted: boolean;
  rmvGrade?: Grade | null;
  reviewFromSurah?: number | null;
  reviewFromAyah?: number | null;
  reviewToSurah?: number | null;
  reviewToAyah?: number | null;
  reviewCompleted: boolean;
  reviewGrade?: Grade | null;
  extraReviewFromSurah?: number | null;
  extraReviewFromAyah?: number | null;
  extraReviewToSurah?: number | null;
  extraReviewToAyah?: number | null;
  teacherNotes?: string | null;
}

export interface StudentStats {
  studentId: number;
  totalAyahsMemorized: number;
  totalQuranPercentage: number;
  juzCompleted: number;
  successfulDaysPercent: number;
  currentStreak: number;
  thisMonthSuccessRate: number;
  ayahsMemorizedThisMonth: number;
  ayahsMemorizedThisWeek: number;
  ayahsMemorizedLastWeek: number;
}

export type CalendarDayStatus =
  (typeof CalendarDayStatus)[keyof typeof CalendarDayStatus];

export const CalendarDayStatus = {
  successful: "successful",
  partial: "partial",
  failed: "failed",
  absent: "absent",
} as const;

export interface CalendarDay {
  date: string;
  status: CalendarDayStatus;
}

export interface CalendarData {
  yearMonth: string;
  days: CalendarDay[];
  successfulDays: number;
  totalAttendedDays: number;
  successRate: number;
}

export type TodayStatusStatus =
  (typeof TodayStatusStatus)[keyof typeof TodayStatusStatus];

export const TodayStatusStatus = {
  all_done: "all_done",
  in_progress: "in_progress",
  not_started: "not_started",
} as const;

export interface TodayStatus {
  status: TodayStatusStatus;
  completedTasks: number;
  totalTasks: number;
}

export type DashboardStudentTodayStatus =
  (typeof DashboardStudentTodayStatus)[keyof typeof DashboardStudentTodayStatus];

export const DashboardStudentTodayStatus = {
  all_done: "all_done",
  in_progress: "in_progress",
  not_started: "not_started",
} as const;

export interface DashboardStudent {
  id: number;
  name: string;
  currentSurah: number;
  currentAyah: number;
  todayStatus: DashboardStudentTodayStatus;
  completedTasks: number;
}

export interface PerformerSummary {
  studentId: number;
  name: string;
  successRate: number;
}

export interface ClassStats {
  totalStudents: number;
  averageSuccessRate: number;
  totalAyahsMemorized: number;
  topPerformers: PerformerSummary[];
  needsAttention: PerformerSummary[];
}

export interface SurahInfo {
  number: number;
  name: string;
  nameArabic: string;
  ayahCount: number;
  juz: number;
}

export type ListStudentsParams = {
  /**
   * Filter by active status
   */
  active?: boolean;
};

export type ListEntriesParams = {
  /**
   * Filter by month (YYYY-MM format)
   */
  month?: string;
  limit?: number;
  offset?: number;
};