# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Quran Hifdh Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Teacher login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/session:
    get:
      operationId: getSession
      tags: [auth]
      summary: Get current session
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /students/{studentId}/entries:
    get:
      operationId: listEntries
      tags: [entries]
      summary: List daily entries for a student
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: string
          description: "Filter by month (YYYY-MM format)"
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyEntry"
  /students/{studentId}/entries/{date}:
    get:
      operationId: getEntry
      tags: [entries]
      summary: Get entry for a specific date
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daily entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertEntry
      tags: [entries]
      summary: Create or update entry for a date
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertEntryRequest"
      responses:
        "200":
          description: Entry saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyEntry"
  /students/{studentId}/stats:
    get:
      operationId: getStudentStats
      tags: [stats]
      summary: Get student KPIs
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentStats"
  /students/{studentId}/calendar/{yearMonth}:
    get:
      operationId: getStudentCalendar
      tags: [stats]
      summary: Get monthly calendar data for a student
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
        - name: yearMonth
          in: path
          required: true
          schema:
            type: string
          description: "YYYY-MM format"
      responses:
        "200":
          description: Calendar data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarData"
...
  /students/{studentId}/today-status:
    get:
      operationId: getStudentTodayStatus
      tags: [entries]
      summary: Get today's completion status for a student
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Today status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodayStatus"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginRequest:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    AuthResponse:
      type: object
      properties:
        authenticated:
          type: boolean
      required:
        - authenticated
...
    Grade:
      type: string
      enum:
        - excellent
        - good
        - needs_repeat
        - incomplete
    DailyEntry:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        date:
          type: string
          format: date
        newMemorizationFromSurah:
          type: integer
          nullable: true
        newMemorizationFromAyah:
          type: integer
          nullable: true
        newMemorizationToSurah:
          type: integer
          nullable: true
        newMemorizationToAyah:
          type: integer
          nullable: true
        newMemorizationCompleted:
          type: boolean
        newMemorizationGrade:
          $ref: "#/components/schemas/Grade"
          nullable: true
        rmvFromSurah:
          type: integer
          nullable: true
        rmvFromAyah:
          type: integer
          nullable: true
        rmvToSurah:
          type: integer
          nullable: true
        rmvToAyah:
          type: integer
          nullable: true
        rmvCompleted:
          type: boolean
        rmvGrade:
          $ref: "#/components/schemas/Grade"
          nullable: true
        reviewFromSurah:
          type: integer
          nullable: true
        reviewFromAyah:
          type: integer
          nullable: true
        reviewToSurah:
          type: integer
          nullable: true
        reviewToAyah:
          type: integer
          nullable: true
        reviewCompleted:
          type: boolean
        reviewGrade:
          $ref: "#/components/schemas/Grade"
          nullable: true
        extraReviewFromSurah:
          type: integer
          nullable: true
        extraReviewFromAyah:
          type: integer
          nullable: true
        extraReviewToSurah:
          type: integer
          nullable: true
        extraReviewToAyah:
          type: integer
          nullable: true
        teacherNotes:
          type: string
          nullable: true
        daySuccessful:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - date
        - newMemorizationCompleted
        - rmvCompleted
        - reviewCompleted
        - daySuccessful
        - createdAt
    UpsertEntryRequest:
      type: object
      properties:
        newMemorizationFromSurah:
          type: integer
          nullable: true
        newMemorizationFromAyah:
          type: integer
          nullable: true
        newMemorizationToSurah:
          type: integer
          nullable: true
        newMemorizationToAyah:
          type: integer
          nullable: true
        newMemorizationCompleted:
          type: boolean
        newMemorizationGrade:
          $ref: "#/components/schemas/Grade"
          nullable: true
        rmvFromSurah:
          type: integer
          nullable: true
        rmvFromAyah:
          type: integer
          nullable: true
        rmvToSurah:
          type: integer
          nullable: true
        rmvToAyah:
          type: integer
          nullable: true
        rmvCompleted:
          type: boolean
        rmvGrade:
          $ref: "#/components/schemas/Grade"
          nullable: true
        reviewFromSurah:
          type: integer
          nullable: true
        reviewFromAyah:
          type: integer
          nullable: true
        reviewToSurah:
          type: integer
          nullable: true
        reviewToAyah:
          type: integer
          nullable: true
        reviewCompleted:
          type: boolean
        reviewGrade:
          $ref: "#/components/schemas/Grade"
          nullable: true
        extraReviewFromSurah:
          type: integer
          nullable: true
        extraReviewFromAyah:
          type: integer
          nullable: true
        extraReviewToSurah:
          type: integer
          nullable: true
        extraReviewToAyah:
          type: integer
          nullable: true
        teacherNotes:
          type: string
          nullable: true
      required:
        - newMemorizationCompleted
        - rmvCompleted
        - reviewCompleted
    StudentStats:
      type: object
      properties:
        studentId:
          type: integer
        totalAyahsMemorized:
          type: integer
        totalQuranPercentage:
          type: number
          format: float
        juzCompleted:
          type: integer
        successfulDaysPercent:
          type: number
          format: float
        currentStreak:
          type: integer
        thisMonthSuccessRate:
          type: number
          format: float
        ayahsMemorizedThisMonth:
          type: integer
        ayahsMemorizedThisWeek:
          type: integer
        ayahsMemorizedLastWeek:
          type: integer
      required:
        - studentId
        - totalAyahsMemorized
        - totalQuranPercentage
        - juzCompleted
        - successfulDaysPercent
        - currentStreak
        - thisMonthSuccessRate
        - ayahsMemorizedThisMonth
        - ayahsMemorizedThisWeek
        - ayahsMemorizedLastWeek
    CalendarDay:
      type: object
      properties:
        date:
          type: string
          format: date
        status:
          type: string
          enum:
            - successful
            - partial
            - failed
            - absent
      required:
        - date
        - status
    CalendarData:
      type: object
      properties:
        yearMonth:
          type: string
        days:
          type: array
          items:
            $ref: "#/components/schemas/CalendarDay"
        successfulDays:
          type: integer
        totalAttendedDays:
          type: integer
        successRate:
          type: number
          format: float
      required:
        - yearMonth
        - days
        - successfulDays
        - totalAttendedDays
        - successRate
    TodayStatus:
      type: object
      properties:
        status:
          type: string
          enum:
            - all_done
            - in_progress
            - not_started
        completedTasks:
          type: integer
        totalTasks:
          type: integer
      required:
        - status
        - completedTasks
        - totalTasks
...