# lib/api-zod/src/generated/types/calendarDay.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Hifdh Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { CalendarDayStatus } from "./calendarDayStatus";

export interface CalendarDay {
  date: Date;
  status: CalendarDayStatus;
}