# lib/api-zod/src/generated/types/grade.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Hifdh Tracker API
 * OpenAPI spec version: 0.1.0
 */

export type Grade = (typeof Grade)[keyof typeof Grade];

export const Grade = {
  excellent: "excellent",
  good: "good",
  needs_repeat: "needs_repeat",
  incomplete: "incomplete",
} as const;