# lib/api-zod/src/generated/types/studentStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Hifdh Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentStats {
  studentId: number;
  totalAyahsMemorized: number;
  totalQuranPercentage: number;
  juzCompleted: number;
  successfulDaysPercent: number;
  currentStreak: number;
  thisMonthSuccessRate: number;
  ayahsMemorizedThisMonth: number;
  ayahsMemorizedThisWeek: number;
  ayahsMemorizedLastWeek: number;
}