# lib/api-zod/src/generated/types/todayStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Hifdh Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { TodayStatusStatus } from "./todayStatusStatus";

export interface TodayStatus {
  status: TodayStatusStatus;
  completedTasks: number;
  totalTasks: number;
}